DBZ-5079: fix: fix the unit test
